==> 2_find_SNPs.ipynb <==
## Important

`run_raxml.sh` in `Genome_tree` to build tree this notebook.

import os
import json
from collections import Counter, defaultdict

import pandas as pd
from Bio import SeqIO, AlignIO, Phylo
from Bio.Seq import Seq
from Bio.SeqRecord import SeqRecord
from treetime import TreeAnc

genomeTreeDir = "Genome_tree"
snpDir = "SNPs"
dataDir = "Data"

if not os.path.exists(snpDir):
    os.mkdir(snpDir)

ambiguity = {
    "Y": ["C", "T"], "R": ["A", "G"], "W": ["A", "T"],
    "S": ["G", "C"], "K": ["T", "G"], "M": ["C", "A"],
    "D": ["A", "G", "T"], "V": ["A", "C", "G"],
    "H": ["A", "C", "T"], "B": ["C", "G", "T"]
}

with open(os.path.join(genomeTreeDir, "outgroup.txt")) as f:
    for row in f:
        exclude = row.strip("\n")

tree = Phylo.read(os.path.join(genomeTreeDir, "RAxML_bestTree.GTRGAMMA"), "newick")
alignedCDS = AlignIO.read(os.path.join(genomeTreeDir, "aligned_cds.fasta"), "fasta")

treeAnc = TreeAnc(tree = tree, aln = alignedCDS, gtr="Jukes-Cantor", verbose=False)
treeAnc.infer_ancestral_sequences(infer_gtr=True, marginal=True)
alignedAnc = SeqRecord(Seq(''.join(treeAnc.tree.root.sequence)), "ancestral")

alignedCDS.add_sequence("ancestral", str(alignedAnc.seq))
AlignIO.write(alignedCDS, os.path.join(snpDir, "ancestral.fasta"), "fasta")

info = pd.read_csv(os.path.join(dataDir, "info.csv"))
info.columns.values[list(info.columns.values).index("Location")] = "City"
info.columns.values[list(info.columns.values).index("Area")] = "Province"

meta = pd.read_csv(os.path.join(dataDir, "original.csv"))
meta.columns.values[list(meta.columns.values).index("Accession ID")] = "Accession ID"

with open(os.path.join(snpDir, "cds_product.json")) as f:
    cdsAnno = json.load(f)

snpPosIndex = []
for i in range(alignedCDS.get_alignment_length()):
    aaSum = Counter(alignedCDS[:, i])
    if len(aaSum) > 1:
        snpPosIndex.append(i)

allSNP = pd.DataFrame(columns=(
    "Pos", "SNP", "Ref", 
    "Genome pos", "Start index",
    "Product", "Product ID", "Accession ID"
))

for record in alignedCDS:
    if record.id != exclude:
        for index in snpPosIndex:
            ref = alignedAnc[index].upper()
            n = record[index].upper()
            if n != 'N' and n != '-' and ref != n:
                if n in ambiguity:
                    if ref in ambiguity[n]:
                        continue
                pos = index + 1
                posAnno = cdsAnno[str(pos)]
                row = pd.Series(data={
                    "Pos": pos, 
                    "SNP": n,
                    "Ref": ref,
                    "Genome pos": posAnno["genome_pos"],
                    "Start index": posAnno["start_index"],
                    "Product": posAnno["product_name"],
                    "Product ID": posAnno["product_id"],
                    "Accession ID": record.id
                })
                allSNP = allSNP.append(row, ignore_index=True)
                
allSNP = pd.merge(allSNP, info, on="Accession ID", how="left", suffixes=("", "_y"))
allSNP = pd.merge(allSNP, meta, on="Accession ID", how="left", suffixes=("", "_y"))
allSNP = allSNP.sort_values(["Pos", "SNP"])
allSNP.to_csv(os.path.join(snpDir, "all.csv"), index=False)

hubeiSNP = allSNP[allSNP["Province"] == "Hubei"][["Pos", "SNP"]].drop_duplicates()

provinceSNP = pd.DataFrame(columns=allSNP.columns)
citySNP = pd.DataFrame(columns=allSNP.columns)

for i in allSNP[allSNP["Province"] != "Hubei"].groupby(["Pos", "Ref"]):
    inHubei = any((hubeiSNP["Pos"] == i[0][0]) & (hubeiSNP["SNP"] == i[0][1]))
    isProvince = len(i[1]["Province"].drop_duplicates()) >= 2
    isCity = len(i[1]["City"].drop_duplicates()) >= 2
    if not inHubei:
        if isProvince:
            provinceSNP = provinceSNP.append(i[1], ignore_index=True)
        if isCity:
            citySNP = citySNP.append(i[1], ignore_index=True)

provinceSNP.to_csv(os.path.join(snpDir, "province.csv"), index=False)
citySNP.to_csv(os.path.join(snpDir, "city.csv"), index=False)

nonSNP = defaultdict(list)

for _, row in allSNP[["Ref", "Pos", "SNP"]].drop_duplicates().iterrows():
    mut = ''.join((row["Ref"], str(row["Pos"]), row["SNP"]))
    nonSNP[mut].extend(allSNP.loc[(allSNP["Pos"] == row["Pos"]) & (allSNP["SNP"] == row["SNP"]), "Accession ID"].values)

nonSNPdir = os.path.join(snpDir, "nonSNP")
if not os.path.exists(nonSNPdir):
    os.mkdir(nonSNPdir)

for mut, accession in nonSNP.items():
    meta[meta["Accession ID"].isin(accession)].to_csv(os.path.join(snpDir, "nonSNP", mut + ".csv"), index=False)

seqs = SeqIO.index(os.path.join(genomeTreeDir, "aligned_cds.fasta"), "fasta")

ancIdentical = pd.read_csv(os.path.join(dataDir, "original.csv"))
ancIdentical = ancIdentical[ancIdentical["Accession ID"].isin(list(seqs.keys()))]
ancIdentical = ancIdentical[~ancIdentical["Accession ID"].isin(allSNP["Accession ID"])]
ancIdentical.to_csv(os.path.join(snpDir, "identical.csv"), index=False)
